+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes:
 - Removed default `default_max_age` from session claim base classes
 - Added a 5 minute `default_max_age` to UserRoleClaim, PermissionClaim and EmailVerificationClaim
+- Fix Repetition of root_path in supertokens mididdlware for fastapi [#230](https://github.com/supertokens/supertokens-python/issues/230)
 
 ## [0.11.1] - 2022-09-28
 ### Changes: